requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
pydantic>=2.0.0
pyyaml>=6.0.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0

